Move StatusManager to pkg/controller
