+++ pkg/controller/networkconfig/networkconfig_controller.go
@@ -11,9 +11,9 @@ import (
 	configv1 "github.com/openshift/api/config/v1"
 	networkoperatorv1 "github.com/openshift/cluster-network-operator/pkg/apis/networkoperator/v1"
 	"github.com/openshift/cluster-network-operator/pkg/apply"
+	"github.com/openshift/cluster-network-operator/pkg/controller/statusmanager"
 	"github.com/openshift/cluster-network-operator/pkg/names"
 	"github.com/openshift/cluster-network-operator/pkg/network"
-	"github.com/openshift/cluster-network-operator/pkg/util/clusteroperator"
 
 	appsv1 "k8s.io/api/apps/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -41,12 +41,12 @@ var ManifestPath = "./bindata"
 
 // Add creates a new NetworkConfig Controller and adds it to the Manager. The Manager will set fields on the Controller
 // and Start it when the Manager is Started.
-func Add(mgr manager.Manager, status *clusteroperator.StatusManager) error {
+func Add(mgr manager.Manager, status *statusmanager.StatusManager) error {
 	return add(mgr, newReconciler(mgr, status))
 }
 
 // newReconciler returns a new reconcile.Reconciler
-func newReconciler(mgr manager.Manager, status *clusteroperator.StatusManager) *ReconcileNetworkConfig {
+func newReconciler(mgr manager.Manager, status *statusmanager.StatusManager) *ReconcileNetworkConfig {
 	configv1.Install(mgr.GetScheme())
 	return &ReconcileNetworkConfig{
 		client: mgr.GetClient(),
@@ -96,7 +96,7 @@ type ReconcileNetworkConfig struct {
 	// that reads objects from the cache and writes to the apiserver
 	client client.Client
 	scheme *runtime.Scheme
-	status *clusteroperator.StatusManager
+	status *statusmanager.StatusManager
 
 	podReconciler *ReconcilePods
 }
